refactor(#9) databaseまわりのディレクトリ構成を修正
