+++ backend/.gitignore
@@ -1,6 +1,6 @@
 api
-ent/*/*
-ent/*.go
-!ent/schema/*
-!ent/generate.go
+database/ent/*/*
+database/ent/*.go
+!database/ent/schema/*
+!database/ent/generate.go
 .migrate.sh
